+++ src/libs/stat.h
@@ -5,9 +5,9 @@
 
 struct stat {
     uint32_t st_mode;                   // protection mode and file type
-    size_t st_nlinks;                   // number of hard links
-    size_t st_blocks;                   // number of blocks file is using
-    size_t st_size;                     // file size (bytes)
+    uint32_t st_nlinks;                 // number of hard links
+    uint32_t st_blocks;                 // number of blocks file is using
+    uint32_t st_size;                   // file size (bytes)
 };
 
 #define S_IFMT          070000          // mask for type of file
